Add counter.ex
The first line instructs Phoenix to use the Phoenix.LiveView behaviour.

This loads just means that we will need to implement certain functions for our live view to work.

The first function is mount/3 which, as it's name suggests, mounts the module with the _params, _session and socket arguments.

In our case we are ignoring the _params and _session, hence the underscore prepended to the parameters. If we were using sessions for user management, we would need to check the session variable, but in this simple counter example we just ignore it.

mount/3 returns a tuple: {:ok, assign(socket, :val, 0)} which uses the assign/3 function to assign the :val key a value of 0 on the socket. That just means the socket will now have a :val which is initialised to 0.

The second function is handle_event/3 which handles the incoming events received. In the case of the first declaration of handle_event("inc", _, socket) it pattern matches the string "inc" and increments the counter.

The third function is almost identical to the one above, the key difference is that it decrements the :val.

Finally the forth function render/1 receives the assigns argument which contains the :val state and renders the template using the @val template variable.
The render/1 function renders the template included in the function. The ~L""" syntax just means "treat this multiline string as a LiveView template" The ~L sigil is a macro included when the use Phoenix.LiveView is invoked at the top of the file.
LiveView will invoke the mount/3 function and will pass the result of mount/3 to render/1 behind the scenes.
Each time an update happens (e.g: handle_event/3) the render/1 function will be executed and updated data (in our case the :val count) is sent to the client.